+++ NEWS.md
@@ -1,3 +1,5 @@
+# ddml (development version)
+
 # ddml 0.2.2
 
 * Changes ``ddml:ols()`` default to ``const=TRUE``.